requirements.txt
aws-cdk.core==1.124.0
aws-cdk.aws-stepfunctions==1.124.0
aws-cdk.aws-stepfunctions-tasks==1.124.0
aws-cdk.aws-eks==1.124.0
aws-cdk.aws-iam==1.124.0
aws-cdk.aws-ec2==1.124.0
aws-cdk.aws-codepipeline==1.124.0
aws-cdk.aws-batch==1.124.0
aws-cdk.aws-codedeploy==1.124.0
aws-cdk.aws-codepipeline-actions==1.124.0
aws-cdk.aws-events-targets==1.124.0
aws-cdk.aws-glue==1.124.0
aws-cdk.aws-kinesisfirehose==1.124.0
aws-cdk.aws-servicecatalog==1.124.0
PyYAML==5.3.1
pydantic==1.6.2
requests==2.24.0
